feat: data collection logic and UI: get screenshots and csv of events that passed stage 1 and have a valid fit. batch different files. just a few clicks from main_window_home. magnitudes are now stored as float64 instead of float32 to avoid overflow due to corruption.

+++ src/core/click_pipeline_v5.py
@@ -738,7 +738,9 @@ def compute_fft_energy(fft_magnitudes: np.ndarray) -> float:
         return 0.0
     # Use only as many bins as are present (guard against truncated frames)
     k = min(len(fft_magnitudes), _K_BINS)
-    return float(np.mean(fft_magnitudes[:k] ** 2))
+    # Cast to float64 before squaring: float32 overflows for large (but physically
+    # plausible) values that arrive from corrupted or high-amplitude frames.
+    return float(np.mean(fft_magnitudes[:k].astype(np.float64) ** 2))
 
 
 ### STAGES ###
